review comments  - port team_update_service to team/update_service  - refactor
