+++ app/services/team/update_service.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+module Team
+  class UpdateService < ApplicationService
+    attr_reader :user_params, :current_company, :new_role, :user, :current_role
+
+    def initialize(user_params:, current_company:, new_role:, user:)
+      @user = user
+      @user_params = user_params
+      @current_company = current_company
+      @current_role = user.roles.find_by(resource: current_company)
+      @new_role = new_role
+    end
+
+    def process
+      User.transaction do
+        # skip confirmation email as changes are made by admin and it dosen't change email
+        user.skip_reconfirmation!
+        user.update!(user_params)
+        update_company_user_role
+      end
+    end
+
+    private
+
+      def update_company_user_role
+        if current_role.present?
+          user.remove_role(current_role.name.to_sym, current_company)
+        end
+
+        user.add_role(new_role.downcase.to_sym, current_company)
+      end
+  end
+end
